Move C++ `ExternType` to its own file
