+++ crates/c-api/include/wasmtime/types.hh
@@ -2,6 +2,7 @@
 #define WASMTIME_TYPES_HH
 
 #include <wasmtime/types/export.hh>
+#include <wasmtime/types/extern.hh>
 #include <wasmtime/types/func.hh>
 #include <wasmtime/types/global.hh>
 #include <wasmtime/types/import.hh>
